+++ privaite/pii/recognizer_fr_date.py
@@ -39,7 +39,7 @@ class FrenchDateRecognizer(EntityRecognizer):
         pass
 
     def analyze(
-        self, text: str, entities: list[str], nlp_artifacts: NlpArtifacts = None
+        self, text: str, entities: list[str], nlp_artifacts: NlpArtifacts | None = None
     ) -> list[RecognizerResult]:
         results = []